* dev: (23 commits)
  Fixing plugin bugs when using longdouble=float128
  Kill off some more warnings
  PTA model confusing pmdec and pmra.
  Additional fixes for long double vs long double in predictors.
  Various fixes for T2predictors to be independent of TKlongdouble
  More unit tests for float128
  Remove warnings
  Fix some scanf for long double
  Capacity for printing float128 - need to use ld_printf.
  Begin conversion to be able to use float128
  Added in pulse numbering
  Options for averaging residuals
  Average data options
  added in averaging residual options
  Use LAPACK least-squares if available.
  gitignore
  Added program to get the period from a police
  Added some trivial tests
  Added google test framework code
  add option to print pulsar frame time in general2
  ...

Conflicts:
	.gitignore
	Makefile.am
	doFit.C
	doconfigure
	plugin/plk_plug.C

+++ ELL1model.C
@@ -84,7 +84,7 @@ double ELL1model(pulsar *psr,int p,int ipos,int param)
     {
       displayMsg(1,"BIN1","SIN I > 1.0, setting to 1: should probably use DDS model","",psr[p].noWarnings);
       si = 1.0;
-      psr[p].param[param_sini].val[0] = 1.0L;
+      psr[p].param[param_sini].val[0] = longdouble(1.0);
     }
 
   x0    = psr[p].param[param_a1].val[0];
