+++ secularMotion.C
@@ -50,8 +50,8 @@ void secularMotion(pulsar *psr,int npsr)
 	  if (psr[0].param[param_dshk].paramSet[0]==1 && 
 	      (psr[0].param[param_pmra].paramSet[0]==1 || psr[0].param[param_pmdec].paramSet[0]==1))
 	    {
-	      t0 = (psr[p].obsn[i].bat - psr[p].param[param_posepoch].val[0])*86400.0L;	      
-	      psr[p].obsn[i].shklovskii = t0*t0/2.0L/299792458.0L*
+	      t0 = (psr[p].obsn[i].bat - psr[p].param[param_posepoch].val[0])*longdouble(86400.0);	      
+	      psr[p].obsn[i].shklovskii = t0*t0/longdouble(2.0)/longdouble(299792458.0)*
 		(getParameterValue(&psr[p],param_dshk,0)*kpc2m)*
 		(psr[p].param[param_pmra].val[0]*psr[p].param[param_pmra].val[0]*
 		 mas_yr2rad_s*mas_yr2rad_s+
@@ -61,7 +61,7 @@ void secularMotion(pulsar *psr,int npsr)
 	  else
 	    psr[p].obsn[i].shklovskii = 0.0;
 	  
-	  psr[p].obsn[i].bbat = psr[p].obsn[i].bat - (psr[p].obsn[i].shklovskii)/86400.0L;
+	  psr[p].obsn[i].bbat = psr[p].obsn[i].bat - (psr[p].obsn[i].shklovskii)/longdouble(86400.0);
 	}
     }
 }
